Update Proposal - Task 1 (Alan)
I have included the write up for the Motivations and Purpose of our application/dashboard. The write up includes information relating to information regarding the reasoning for creating the application which pertain to helping small businesses/private sellers/entrepreneurs understand the market of the 2nd hand market in the United states - based off of 2nd hand car sales from the United States.